+++ pt_os_web_portal/os_updater/__init__.py
@@ -1,0 +1,1 @@
+from .updater import OSUpdater
