+++ pt_os_web_portal/os_updater/types.py
@@ -1,0 +1,14 @@
+from enum import Enum, auto
+
+
+class MessageType(Enum):
+    ERROR = auto()
+    START = auto()
+    STATUS = auto()
+    FINISH = auto()
+
+
+class EventNames(Enum):
+    OS_UPGRADE = auto()
+    OS_PREPARE_UPGRADE = auto()
+    SIZE = auto()
